+++ Quasardb/Table/Reader/QdbCell.cs
@@ -97,7 +97,7 @@ namespace Quasardb.TimeSeries.Reader
                 if (err == qdb_error.qdb_e_element_not_found)
                     return null;
                 QdbExceptionThrower.ThrowIfNeeded(err, alias: _alias, column: _column.name);
-                return Helper.GetBytes(content, (int)length);
+                return Helper.GetBytes(content, length);
             }
         }
 
@@ -138,8 +138,8 @@ namespace Quasardb.TimeSeries.Reader
                 if (err == qdb_error.qdb_e_element_not_found)
                     return null;
                 QdbExceptionThrower.ThrowIfNeeded(err, alias: _alias, column: _column.name);
-                // TODO: limited to 32-bit
-                var value = Helper.GetBytes(content, (int)length);
+
+                var value = Helper.GetBytes(content, length);
                 if (value == null)
                     return null;
                 return System.Text.Encoding.UTF8.GetString(value);
